airflow/requirements.txt
# Airflow provider
apache-airflow-providers-postgres>=5.10.0

# ETL Shared dependencies (MUST MATCH root)
psycopg[binary]==3.2.1
pika==1.3.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2

# Selenium + Webdriver
selenium==4.25.0
webdriver-manager==4.0.2
undetected-chromedriver==3.5.5
